Module "screen" should be loaded before other modules that use graphics
Fixes issue 8.

Finally! In OpenGL version, this bug was seen because Font was loaded
before screen. Since Font initializes the default font as a texture, it
needs some OpenGL stuff to be correctly set up. screen does this, so it
should be loaded first. Now default font works in OpenGL yay!