GLCanvas: add a zero at the end of the array when creating the canvas.
On Windows, this was not necessary (and is still untested).
On Linux, I get a segfault immediately when the 0 isn't there.

Linux version still segfaults, but I have still to know if it's my 
version of wxLua/luaglut that's faulty, or anything MLS is doing that
crashes the OpenGL implementation and/or Gfx driver (AMD fglrx 9.11).