No more crash on Linux in OpenGL mode when wxGLCanvas is available.
Aaaaaargh! All the crashes I was seeing were because I forgot to invert
the order of the conditions in screen._switchOffscreen().
wxGLCanvas should be tried first, then SDL as a second choice!
As a result, wxGLCanvas/Context were used to set up OpenGL, but SDL was
use to (try to) swap the buffers => crash!