Change (again!) the way we handle window/canvas resizing.
The thing is, despite setting a MinSize of 256x384 (=1x DS screens) on
the GLCanvas, it didn't prevent Windows from resizing the *window* 
below this size, and the contained canvas happily resized too :( .

So the sizer didn't do its job, it seems it's not able to stop Windows
from resizing, well... windows.

So I had to add a SetSizeHints() to the window after it's first created
and its content fits the 1x screens' size (well, more precisely, I call
SetSizerAndFit(), which in turn calls SetSizeHints() on the window).

It seems to force Windows to behave this time. The main window can't be
resized below its initial size.