No more seams when using scrollmaps and the window has been resized.
Fixes issue 17.

Don't know it it's exactly what's called "texture bleeding", as it only
happens when the "zoom factor" for the window isn't an integer.

Furthermore, I'm using GL_NEAREST for interpolation, so this kind of
artifact should be minimal.

Anyway, I now avoid taking values 0 and 1 for texel coords, I use 0.001
and 0.999, and the glitches seem to have disappeared, at any zoom
factor.

We'll see if the fix causes other problems...