Alternative method to display scaled screens on Linux (non-OpenGL mode).
In the previous commit, I stated that resizing the window in non-OpenGL
mode was a bad idea because it's dog slow. Well it turns out that it's
actually very fast on Windows XP (tested in VirtualBox). I used 
DC:SetUserScale(), which proves to be dog slow on Linux/GTK though.

In the present commit, I use another method on Linux. An image is 
created from the current offscreen bitmap, then rescaled to the size of
the (resized) screens, then converted back to a bitmap, which is finally
drawn on the DC. I thought this wouldn't make any difference, since 
conversions between Image and Bitmap tend to be real slow as well. 
But it seems that whatever Linux/GTK version is doing internally when 
DC:SetUserScale() is used, is even slower.

I'll have to test whether Windows support this method as well (otherwise
I'll keep using DC:SetUserScale()), but most of all, which of the two 
methods works reasonably well on Mac.