First support for OpenGL on Mac OS X.
For this, I had to use a deprecated version of wxGLCanvas' constructor,
that creates the wxGLContext itself, since wxGLContext doesn't seem to
be present on my version of wxLua on Mac. So I commented the explicit
creation of the context, too. I have to test this version on Windows 
and Linux again. If it doesn't work, I'll have to create the canvas 
conditionnally.

The current problems in Mac version of OpenGL mode:
 - no display until the window is first manually resized
 - current frame isn't seen whenever the script is paused (black 
   background instead)
 - message about a bad malloc at launch