OpenGL is set up differently on Mac OS X and Windows/Linux.
The version of wxWidgets I provide for Mac is older, and doesn't use 
separate constructors to create the GL canvas and the GL context.
It only has the canvas constructor that automatically and implicitely
creates the context, so that's what we use.

On Windows/Linux, on the other hand, we create the canvas first, then
we have to create the context explicitely.