Loading and displaying textures in OpenGL is now A LOT less retarded.
Man, you can tell I was struggling for even displaying something that
works in OpenGL, when I began. I loaded the texture upside down, then
thought it was funny to display it correctly using glScale(1, -1, 1)
in texture matrix!
It's REALLY easier to load the texture the right way, then *simply*
display it!!!
The good thing is that glScale() wasn't working with texture rectangle,
but now the latter displays things correctly.
Is it the end of the NPOT textures problems for all cards??? I hope so.