screen: WIP: implement "internal" ML functions.
These functions exist in a few modules, and are exported by ML C code,
because the Lua libs that are launched at boot use them.

This is a first step, and because some of ML libs are implemented in 
Lua instead of C, some of the MLS libs will eventually disappear.

The goal is that the "internal" (low-level) functions in MLS behave as
much as possible like ML, so the libs of ML work with MLS.

One of the next problem will be to change all MLS code that use
SCREEN_UP and SCREEN_DOWN, because I set them to 0 and 192, which made
screen functions shorter and faster, but in ML their values are 1 and
0, so no more shortcuts in all my gfx functions :(