Changed values of SCREEN_UP and SCREEN_DOWN.
They go from 0,192 to 1,0.

This forced me to change many functions, because in the old version I
relied on the fact that the number of the screen directly matched the 
screen Y-offset from the top (the bottom screen begins at y=192).

Now I have to use a table to match the screen number with its Y-offset.
(I could have done it with a function, but I guess a table access is 
faster than a function call/return ?)