Change compile system to prevent class/module names clashes.
With the previous version, compiled MLS would have crashed or misbehaved
if a table in user script was simply called, for example, Logger, or 
Observable, or Gui, or any of the classes used by MLS.

That's because my classes, when compiled, would be named with the 
filename minus the extension. Now, every class/module is named in the
compiled version after it complete location, that is clp_Class for 
clp.Class, or clp_mls_Config for clp.mls.Config.

Right now, only the Bash compile script has been changed, the Lua 
version still needs to be converted.