screen.drawTextBox() currently assumes there's always a DC available.
We should normally need a DC in wx mode only, especially since r428, 
where the DCs aren't created anymore in OpenGL mode.

The quick fix I applied here may be needed in other methods that aren't
overridden by screen (gl). Whether it is the case or not, I should try
to find a more elegant, "inheritance-aware", way to fix this.