ScriptManager: replace pcall() with a custom version in user scripts.
This is done to avoid the "attempt to yield across metamethod/C-call 
boundary" error.

Actually this will replace pcall() not only in user scripts, but also
in MLS, should it need pcall(), so I should find a more subtle way to
fix this (replace pcall() calls directly in the user script?).

At the moment, MLS calls pcall() only in ModuleManager, and the call 
occurs before the replacement is made, so it should be ok.

Also, the custom version doesn't help the user script, since it always
return true (success) + return values from the called function => pcall
doesn't "catch" anything as it is supposed to.