Change default behaviour/values for boot scripts and libs emulation.
If a user script is not given on the command line, then the boot script
will automatically be loaded. This is the only way to launch a boot 
script.

The default boot script is /lua/libs/libs.lua. It can be changed in the
config file.

The default value for libs emulation when it's not set in the config 
file depends whether a boot script has been launched or not.
If a boot script is used, then libs emulation is disabled by default.
If a user script is launched, then libs emulation is enabled by default.
On the other hand, if emulate_libs is explicitely set in the config 
file, then its value is used.