TEST: prevent coroutine.yield() when inside a pcall().
The only yield() is inside Controls.read(), so pcall's in user scripts 
shouldn't be a problem if they don't contain Controls.read().

Unfortunately, libs.lua loads shell.lua inside a pcall(), and shell 
is a giant loop which contains a Controls.read().

So it never gives control back...

I'm not sure there's a way to fix this :( (except rewriting the libs).