Always set magenta as transparent color in images.
Even when an image already has a transparent color that is different 
from magenta, we apparently have to se the latter as a transparent color
too.

In wx mode, this was "automatic" in some cases (at least for ScrollMaps)
but in OpenGL mode, we have to do it explicitly.

Fixes issue 38.