OpenGL: images with real alpha channel (e.g. PNGs) were not loaded correctly.
You can see this in Pokemon Battle, where the font is a PNG image with alpha
channel for the background, and black pixels for the characters.

Since only images with masks (e.g. GIFs) were searched for "transparency", 
the alpha channel in  this font/image was not kept, so the characters were
black blocks (no alpha = all pixels opaque).

This has not been extensively tested, I hope it doesn't cause problems 
elsewhere.