Add some "clp" lib classes to work with basic geometry.
Currently, there are 3 classes: Point, Vector, and Plane.

I needed this to deal with clipping planes in OpenGL, and since it could be
useful for something else (who knows?), I built it as a set of classes 
instead of some method integrated to MLS.

I'm not really satisfied with the way multiple "signatures" are implemented
(e.g. for Point.add...() or Vector.new...()), and I still have to understand
(and then implement) fully how I could create planes more easily while 
predicting its "OpenGL" front/back side (important for clipping), but it'll
do for now.