Plane: add "pseudo-constructor" newFromPointAndNormalVector().
Actually, the code inside it was already in ctr().

Now it is in a separate method that you can call directly, BUT that method is
still called in ctr(), as the default way to build a Plane.

I could decide later to make ctr() "guess" which exact pseudo-constructor
to call, depending on the parameters passed.