screen.drawTextBox() was broken in OpenGL mode.
There were issues with clipping, which caused the text not to show depending
on what other drawing functions were called before drawTextBox(), and on which
screen.

The current fix is a ugly hack, and drawTextBox() should be cleaned up to work
in a generic way, in both OpenGL and wx mode.

Meanwhile, this revision should fix issue 41.