Simplify setClippingRegion() in screen (gl).
There's no need to create Point, Vector and Plane objects to compute the planes
equations.

We can define the unit normals to the planes ourselves, since we know the four
planes needed to form the clipping box are // to the X or Y axis.

All we have to do then is to use directly the x's and y's of the box corners as
the "d" variable, with the right sign, so the right "side" of the plane is 
clipped.

So, no more complicated object creations and equation computing.

(the caching of the previous clipping box is still needed to not bring 
KnightmareDS to a crawl in OpenGL, though, so maybe I should check the way
calls to clipping are done and reduce them?)