Sprite: first try at start/stop/resetAnimation() implementation.
I just tested it with ZeldaLua, with which MLS had problems before.
When using the sword, the animation got caught in an endless loop because
of the lack of resetAnimation() implementation.

It seems to work ok now.

Also, the homebrews that use only playAnimation() seem to work just as 
before.

Fixes issue 6.