Class: inheritance can be set up after class creation.
Inheritance setup has been move to a method, which can be called after the 
Class object has been created.

Another change is that you can now specify the name of the local variable you 
use to reference your "class" in its definition file. I personally use M (for 
"module") all the time.
This change won't be a big need for me, but it could help me fix the compile
script :)