compile (lua) up to date for the lastest Class' changes. WARNING, READ MODE...
I just discovered that the whole new Class' inheritance system, based on the 
replacement of upvalues directly inside "copies" of methods (not references), 
was NOT WORKING WHEN COMPILED LUA IS USED !!!

For now, you can "compile" MLS into one file, but you must the --plain option,
otherwise inheritance won't work.

This is probably due to debug symbols being stripped from the lua files when 
they're compiled.

This is too dumb, I should've seen this one coming. I'll have a look and see
if it can be fixed at all :-/ .