Compiled version of MLS. READ MORE...
It is confirmed that the new Class' inheritance system DOES NOT WORK AT ALL
WITH COMPILED LUA CODE *IF DEBUG SYMBOLS HAVE BEEN STRIPPED*.

So the version compiled here had the debug symbols. There is no option right
now to compile and keep the debug symbols is MLS compile scripts, it has been 
done by hand (maybe I'll add an option later, or make it the default).

Also, it temporarily disable the debug window (not finished).
Please note that in the source version of MLS, the debug window still shows
on launch.