Class: add an alternative way for cloning methods between parent/child classes.
This could be called the "global" way, as it doesn't assume that classes are
declared locally in their file ("local M = Class.new() ... return M"), but 
rather that they are global variables (e.g. stored in "_G['myClass']").

So the way inheritance (with LSB support) is implemented, by cloning methods
from the parent class to the child class, *with references to the class* inside
the methods changed from parent to child, is not the same.
In the "local" version, we rely on the debug symbols being available, and we 
replace upvalues in the cloned methods.
In the "global" version, we change the environment of the cloned methods so 
that references to the parent classes (by index name on the "global" table)
are now references to the child class.

The new "global" method is not enabled by default. It can be by calling  
Class.enableGlobalClasses().

First (uncommited) tests and changes to the compile script to use this method
DO NOT WORK RIGHT NOW, so this is considered a WIP.