ModuleManager: separate modules and modules *names* tables.
There's a default table of modules *names* specifying which modules should be
loaded.

Then, when a module with a given name (and maybe an optional prefix) is found
and successfully loaded, it is stored as an object.

Before, the names *and* the loaded module were in the same table 
(self._modules), the former being string values with numeric keys, the latter
being objects (=tables) with string keys (the name).

This was confusing, so there are now 2 table members in ModuleManager: 
_moduleNames and _modules. They're not the same!