Logger: add a "critical" level (can be any of the existing levels).
All messages equal or above the critical level will get logged, even if the 
category of the message is not "followed" by the logger.

By default, the critical level is set to ERROR.

This prevents the error (and fatal) messages from being hidden when the 
category of the error is not currently followed by MLS (e.g. for tests).

Note that there isn't a setter for this feature right now, only a default.

Fixes issue 50.