ModuleManager: simplify the use and the code related to prefixes.
Code that change the prefixes was all over the place. There was 
self._prefixes, set when the manager is created.
Then, on multiple methods, there was a "prefixes" parameter. And the prefixes
were appended an empty string too many times, also, in _loadModule() and 
_registerCompiledModule().

I chose to remove the prefixes parameters in those methods. Now the prefixes
are set on ModuleManager creation, and you can add some with addPrefix(), 
that's all. They are stored in self._prefixes, and the methods use that.

Also, the empty prefix is only added in loadModules(), and never twice.

Fixes issue 45.