luac.lua: the default output filename is now <input filename>.compiled
This is for coherence with lua AIO and its "-c" option for compiling. 
The default output filename is the name of the input file, with an added 
extension ".compiled".

Since the main reason I created luac.lua is for use with "lua AIO -c", not 
with standard Lua, it makes sense.