compile: do NOT strip symbols by default, + allow Lua AIO to not strip symbols.
The old default, when not using --plain to get a text version of mls.comp.lua,
was to strip symbols from the compiled file. Now the default is to not strip 
symbols.

To force the deletion of symbols from compiled files, you can use the new 
--strip option.

Also, there's no option to not strip symbols when compiling with Lua AIO (all
in one), so we use the fake "compiler", written in pure Lua, which does not
strip symbols, and seems to work with Lua AIO.

I should test the Lua AIO compile with a 64-bit Linux, to see if it works.