ScriptManager: extract environment setting, correct comments, rename a method.
Instead of calling setfenv() directly when we need to mess with some functions
environment, call a custom function that will do it. This might be needed if 
we use a system where a simple setfenv is not enough. If it turns out that 
it's useless, then I'll go back to the old way.