ds_controls: prevent Stylus and Keys tables to ever change their address.
The old way of clearing stylus and keys states were to (inadevertedly) 
create a new table with initial values in it.

The problem with this approach was that the Stylus and Keys tables had a new
address on each module reset, so if any external code had referenced them 
before, it no longer caught stylus and keys state change => the controls did
not work anymore. For example, if emulateLibs is true, the Keys and Status 
tables were copied to the global environment.

The new way of clearing states is to create Status and Keys only once in the 
class, then *copy* the initial values in them on each reset.