Class: inherited classes don't just read from parent env, they *write* too.
If a method in a derived class writes to a global variable, it will be written
in the parent's "original" global environment too (we use a metatable and
__newindex).

This way, if methods from the parent and methods from the child write to the
same global variable, it will work.

Note that writing to (or reading from) global variables is generally bad, but
MLS has to (VRAM, RAM and other constants, NB_FPS).