ScriptManager: change the way MLS methods' environment is replaced.
The new method involves metatables and (new)index metamethods, so it might
be slower than the old way (plain copy of table items) because of the 
necessary "lookup" when an index is not found.

This commit makes the compiled version of MLS work again, *but* the NB_FPS
variable doesn't behave correctly (it's always zero).