ScriptManager: MLS custom version of debug.traceback() was incorrect.
It didn't take into account that the first param (thread) is optional.

After reading the C source for Lua's debug.traceback(), I think its 
processing of the params is rather complicated, so I hope this time I got it
right in the MLS custom version. I only need to increment the level anyway.
The other params can be passed as is.