Class._checkChosenClassSystemIsOk() now works +  debug info removed.
r589 didn't work correctly. It used "nups" info for the function to check that
the debug lib worked (=symbols not stripped from compiled script). But this
is not a good way to check, since nups will always be available, even when the
symbols have been stripped.

On the other hand, the name and value of a specific local variable won't be
available when the symbols have been stripped. So we can use this to do the 
check.

Also, a print statement used for debugging had been left in the commit.