Improve compile scripts (bash and lua versions).
It's now possible to use the --global-classes option, to specify that we want
the classes in MLS to be available as global variables only (on _G).

This is the only working option if you use --strip to strip the symbols from
the compiled script, since local classes won't work in this case.

The default for the script is to declare classes as local variables 
(=upvalues).

Also, there is now a warning displayed when the --plain option is used with 
any option that would only make sense for a compiled script (not a "plain 
text" one). Such options are --luac and --strip.

Finally, an error is displayed and the script is aborted if --strip is used,
but --global-classes is not used. As explained before, --strip and local 
classes (the default behavior) wouldn't work anyway.