ScrollMap (gl) now uses OpenGL display lists, should be faster but READ MORE.
When the scrollmap is loaded, a display list to display it as a whole is
created. So, when draw() is called, we only need to call the display list once
(or more because the scrollmap must be repeated), which is very fast.

On the other hand, like we do in ScrollMap (wx), whenever a tile is changed 
with setTile() , the display list needs to be recompiled. The script 
KnightmareDS uses setTile(), for example.

The current version is good enough for now, but maybe later we'll need to 
speed things up for cases when setTile() is used often.

Possible solutions to investigate then:
  . having one display list per row or per column, and only change the display
    list affected by setTile() when needed
  . having one diplay list per *tile* in the image, and use glCallLists()
    (notice the plural form) to display them in the right order. When 
    setTile() is called, we only need to change the "array" we pass to 
    glCallLists(). Not sure if this will be really fast, or if I even 
    understood correctly how glCallLists is supposed to work... :)