screen (gl): remove dead code related to the (now unused) "fake mouse pointer".
The fake mouse pointer was useful when I was using SDL to create a "floating" 
OpenGL canvas, but had to use the wx main window to get the mouse position =>
I drew the fake mouse pointer with OpenGL at the right position *in* the SDL 
canvas.

I never used that code since I got wx GLCanvas to work, so... bye bye.