Image: fix mirrorH() and mirrorV() behavior.
I managed to get it wrong in two *different* ways (wx and gl), and still it 
wasn't close to how the real uLua works!

Yet it's easy: if the 2nd param is true, then the image will be mirrored, if 
the 2nd parameter is false, then the image will not be mirrored.

Hope I got this right, this time.

Fixes issue 57.