Add a generic Dispatcher class.
As stated, this is very generic, and is only a test right now.

The dispatcher will  dispatch "items", then fetch them again, based on a 
"key" (specific to the item) or an event (yet to be determined).

The items are not written yet, but they'll soon be. So this is a W.I.P.