Add an abstract Item and a test Item for the Dispatcher.
The items classes will be stored in a container subdir.

An item has to extend the AbstractItem class, which provides one method to 
give the "key" for fetching the item, and one method that will get called 
whenever the item is "found" (fetched) by key.

Also, for now, the items names have to be defined inside the Dispatcher, as
they'll not be detected automatically. This would need code that lists files
in a directory etc. Not planned at the moment.

Well, I hope the Dispatcher/Item system is a great piece of well engineered 
code (NOT?), and not a poor over-engineered joke. This would be ironic, 
especially as it's commited on April 1st ;) ;) ;)