Gui: add *COMMENTED* test on files dropped on a surface (THIS IS A TEST).
This should only work on Windows up to wxWidgets 2.8.9, from 2.8.10 it should
exist on Linux and Mac too.

Of course, the test should be done at a more appropriate place, and trigger 
the same functions as the "Open..." menu item.

Also, I think it's only called in OpenGL for now, due to its location in the
code.