Add a TEST application launcher for MLS.
The file mls_launcher.scpt has to be in the same folder as mls.app, so for 
now I always copy them in the root of mls.

Then you can launch mls_launcher.scpt with osascript, and pass it the script
you want to run as argument.

Done like this, it doesn't seem that useful. The trick is that you can compile
this Applescript into an app, and then it can be double-clicked (already 
doable before) *and* it can accept a file dropped on its icon, then pass it
to MLS (I think it's not doable in pure Bash, since we don't get the dropped
file's path).

To compile the script, I use "osacompile -o mls_launcher.app mls_launcher.scpt".

I'd like to find a way not to have 2 apps in parallel, but when I try to exec
the wxLua binary with anything else than Bash, the app doesn't get any input 
on Mac OS X 10.4/Tiger, it doesn't even appear in the dock! :(